feat(): move inputs to their own files
