+++ src/app/views/Diagram/Inspector/Inspector.tsx
@@ -1,11 +1,11 @@
-import React, { ChangeEvent, useEffect, useState } from "react";
+import React from "react";
 import styles from './inspector.module.scss';
-import { Input } from "@mui/material";
 import { dia } from "@clientio/rappid";
-import { colorChangeShapes, getShapeLabelWidth } from "../../../utils/rappid-utils";
-import { MuiColorInput, MuiColorInputValue } from "mui-color-input";
+import { colorChangeShapes } from "../../../utils/rappid-utils";
 import { GlobalShapesTypes } from "../../../types/enums";
 import { MaxLinksInput } from "./MaxLinksInput";
+import { LabelInput } from "./LabelInput";
+import { ColorInput } from "./ColorInput";
 
 interface InspectorProps {
     cellView: dia.CellView
@@ -13,50 +13,7 @@ interface InspectorProps {
 }
 
 const Inspector = ({cellView, graph}: InspectorProps) => {
-    const inspectedElementText = cellView.model.attr("label/text") || "";
     const inspectedGlobalType: GlobalShapesTypes = cellView.model.prop("type");
-
-    const [textValue, setTextValue] = useState(inspectedElementText);
-
-    const findInspectedColor = (inspectedGlobalType: GlobalShapesTypes) => {
-        if (inspectedGlobalType === GlobalShapesTypes.CustomLink) {
-            return cellView.model.attr("line/stroke");
-        } else {
-            return cellView.model.attr("background/fill");
-        }
-    };
-
-    const [color, setColor] = useState(findInspectedColor(inspectedGlobalType));
-
-    const handleLabelChange = (event: ChangeEvent<HTMLInputElement>) => {
-        cellView.model.attr("label/textWrap", {
-            width: getShapeLabelWidth(cellView),
-            height: 20,
-            ellipsis: true
-        })
-
-        setTextValue(event.target.value);
-        cellView.model.attr("label/text", event.target.value);
-    }
-
-    const handleColorChange = (color: MuiColorInputValue) => {
-        setColor(color);
-
-        if (inspectedGlobalType === GlobalShapesTypes.CustomLink) {
-            cellView.model.attr("line/stroke", color);
-        } else {
-            cellView.model.attr("background/fill", color);
-            cellView.model.attr("body/stroke", color);
-        }
-    }
-
-    useEffect(() => {
-        setTextValue(inspectedElementText);
-    }, [inspectedElementText])
-
-    useEffect(() => {
-        setColor(findInspectedColor(inspectedGlobalType));
-    }, [findInspectedColor])
 
     return (
         <div className={styles.inspector}>
@@ -66,8 +23,7 @@ const Inspector = ({cellView, graph}: InspectorProps) => {
                         <div className={styles.inspectorCategoryContainer}>
                             <h4 className={styles.inspectorCategoryHeader}>Label</h4>
                         </div>
-                        <Input className={styles.input} value={textValue}
-                               onChange={handleLabelChange}/>
+                        <LabelInput cellView={cellView}/>
                     </div>
                     <div className={styles.inspectorElement}>
                         <div className={styles.inspectorCategoryContainer}>
@@ -81,7 +37,7 @@ const Inspector = ({cellView, graph}: InspectorProps) => {
                 <div className={styles.inspectorCategoryContainer}>
                     <h4 className={styles.inspectorCategoryHeader}>Color</h4>
                 </div>
-                <MuiColorInput value={color} onChange={handleColorChange}/>
+                <ColorInput cellView={cellView}/>
             </div>}
         </div>
     )
